Added option to natively build the go applications.

+++ build.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+
+while getopts m: option
+do
+case "${option}"
+in
+m) MODE=${OPTARG};;
+esac
+done
+
+
 cd app/worker
 
 echo ""
@@ -9,8 +19,19 @@ echo "## Building executables ##"
 echo "##########################"
 echo ""
 echo ""
-sudo docker run --rm -v "$PWD/../../":/usr/src -w /usr/src/app/worker golang:1.15 go build
-sudo docker run --rm -v "$PWD/../../":/usr/src -w /usr/src/app/deploy golang:1.15 go build
+
+if [ "$MODE" == "native" ];
+then
+    echo "Building natively"
+    echo ""
+    go build
+    cd ../deploy
+    go build
+    cd ../worker
+else
+    sudo docker run --rm -v "$PWD/../../":/usr/src -w /usr/src/app/worker golang:1.15 go build
+    sudo docker run --rm -v "$PWD/../../":/usr/src -w /usr/src/app/deploy golang:1.15 go build
+fi
 
 echo ""
 echo ""
